sdk/py/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stratus"
version = "0.1.0"
description = "Zero-ORM Python SDK for PostgreSQL with connection pooling and type-safe query execution"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Stratus Team"},
]
keywords = [
    "postgresql",
    "database",
    "sdk",
    "orm",
    "type-safe",
    "async",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # asyncpg is the preferred driver (fastest)
    "asyncpg>=0.28.0",
    # psycopg2 is an alternative driver
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/stratusdb/stratus"
Repository = "https://github.com/stratusdb/stratus"
Issues = "https://github.com/stratusdb/stratus/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["stratus*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
